Add files via upload
In analyzing the dataset, “HalloweenData-2018.xslx,” I attempted to create visualizations for my audience to represent the peak time intervals to expect trick-or-treaters, as well as, whether or not the day of the year that Halloween fell on had any impact on the turnout of trick-or-treaters. Within my analysis, I also forecasted the predicted total turnout of trick-or-treaters for the next three calendar years.